Prep for Cargo release
- Split out Anvil support into `bitski-local` package
- Add versions to dependencies
