+++ Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ["bitski", "bitski-provider", "bitski-chain-models"]
+members = ["bitski", "bitski-local", "bitski-provider", "bitski-chain-models"]
