+++ bitski-chain-models/Cargo.toml
@@ -7,6 +7,3 @@ edition = "2021"
 anyhow = "1.0"
 serde = "1.0"
 serde_json = "1.0"
-
-[features]
-local = []
