+++ bitski-provider/Cargo.toml
@@ -7,16 +7,15 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
-bitski-chain-models = {path = "../bitski-chain-models" }
+bitski-chain-models = { version = "0.1.0", path = "../bitski-chain-models" }
 jsonrpc-core = "18"
 oauth2 = "4.0"
 once_cell = "1.14.0"
 reqwest = "0.11"
 serde = "1.0"
 serde_json = "1.0"
-tracing = {version = "0.1", optional = true}
+tracing = { version = "0.1", optional = true }
 web3 = "0.18"
 
 [features]
 tracing = ["dep:tracing"]
-
