+++ bitski/Cargo.toml
@@ -4,11 +4,9 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-anvil = { git = "https://github.com/foundry-rs/foundry", optional = true }
 anyhow = "1.0"
-bitski-chain-models = { path = "../bitski-chain-models" }
-bitski-provider = { path = "../bitski-provider" }
-ethers = { version = "0.17.0", optional = true }
+bitski-chain-models = { version = "0.1.0", path = "../bitski-chain-models" }
+bitski-provider = { version = "0.1.0", path = "../bitski-provider" }
 jsonrpc-core = "18"
 oauth2 = "4.0"
 reqwest = "0.11"
@@ -18,5 +16,4 @@ tracing = { version = "0.1", optional = true }
 web3 = "0.18"
 
 [features]
-local = ["bitski-chain-models/local", "dep:anvil", "dep:ethers"]
 tracing = ["dep:tracing"]
