+++ rust-toolchain.toml
@@ -1,3 +1,3 @@
 [toolchain]
-channel = "1.63"
-components = ["rustfmt", "clippy"]
+channel = "1.64"
+components = ["clippy", "rustfmt"]
